Remove Task and Statement to ...sys.sync
