+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/execution/config/sys/sync/CreatePipeTask.java
@@ -17,26 +17,26 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.execution.config.sys;
+package org.apache.iotdb.db.mpp.plan.execution.config.sys.sync;
 
 import org.apache.iotdb.db.mpp.plan.execution.config.ConfigTaskResult;
 import org.apache.iotdb.db.mpp.plan.execution.config.IConfigTask;
 import org.apache.iotdb.db.mpp.plan.execution.config.executor.IConfigTaskExecutor;
-import org.apache.iotdb.db.mpp.plan.statement.sys.ShowPipeStatement;
+import org.apache.iotdb.db.mpp.plan.statement.sys.sync.CreatePipeStatement;
 
 import com.google.common.util.concurrent.ListenableFuture;
 
-public class ShowPipeTask implements IConfigTask {
+public class CreatePipeTask implements IConfigTask {
 
-  private ShowPipeStatement showPipeStatement;
+  private CreatePipeStatement createPipeStatement;
 
-  public ShowPipeTask(ShowPipeStatement showPipeStatement) {
-    this.showPipeStatement = showPipeStatement;
+  public CreatePipeTask(CreatePipeStatement createPipeStatement) {
+    this.createPipeStatement = createPipeStatement;
   }
 
   @Override
   public ListenableFuture<ConfigTaskResult> execute(IConfigTaskExecutor configTaskExecutor)
       throws InterruptedException {
-    return configTaskExecutor.showPipe();
+    return configTaskExecutor.createPipe();
   }
 }
