+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/execution/config/sys/sync/DropPipeSinkTask.java
@@ -17,34 +17,26 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.statement.sys;
+package org.apache.iotdb.db.mpp.plan.execution.config.sys.sync;
 
-import org.apache.iotdb.db.mpp.plan.analyze.QueryType;
-import org.apache.iotdb.db.mpp.plan.statement.IConfigStatement;
-import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
-import org.apache.iotdb.db.mpp.plan.statement.metadata.ShowStatement;
+import org.apache.iotdb.db.mpp.plan.execution.config.ConfigTaskResult;
+import org.apache.iotdb.db.mpp.plan.execution.config.IConfigTask;
+import org.apache.iotdb.db.mpp.plan.execution.config.executor.IConfigTaskExecutor;
+import org.apache.iotdb.db.mpp.plan.statement.sys.sync.DropPipeSinkStatement;
 
-public class ShowPipeStatement extends ShowStatement implements IConfigStatement {
+import com.google.common.util.concurrent.ListenableFuture;
 
-  private String pipeName;
+public class DropPipeSinkTask implements IConfigTask {
 
-  public String getPipeName() {
-    return pipeName;
-  }
+  private DropPipeSinkStatement dropPipeSinkStatement;
 
-  public void setPipeName(String pipeName) {
-    this.pipeName = pipeName;
+  public DropPipeSinkTask(DropPipeSinkStatement dropPipeSinkStatement) {
+    this.dropPipeSinkStatement = dropPipeSinkStatement;
   }
 
   @Override
-  public QueryType getQueryType() {
-    return QueryType.READ;
-  }
-
-  public ShowPipeStatement() {}
-
-  @Override
-  public <R, C> R accept(StatementVisitor<R, C> visitor, C context) {
-    return visitor.visitShowPipe(this, context);
+  public ListenableFuture<ConfigTaskResult> execute(IConfigTaskExecutor configTaskExecutor)
+      throws InterruptedException {
+    return configTaskExecutor.dropPipeSink();
   }
 }
