+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/execution/config/sys/sync/ShowPipeSinkTask.java
@@ -17,32 +17,26 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.statement.sys;
+package org.apache.iotdb.db.mpp.plan.execution.config.sys.sync;
 
-import org.apache.iotdb.db.mpp.plan.analyze.QueryType;
-import org.apache.iotdb.db.mpp.plan.statement.IConfigStatement;
-import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
-import org.apache.iotdb.db.mpp.plan.statement.metadata.ShowStatement;
+import org.apache.iotdb.db.mpp.plan.execution.config.ConfigTaskResult;
+import org.apache.iotdb.db.mpp.plan.execution.config.IConfigTask;
+import org.apache.iotdb.db.mpp.plan.execution.config.executor.IConfigTaskExecutor;
+import org.apache.iotdb.db.mpp.plan.statement.sys.sync.ShowPipeSinkStatement;
 
-public class ShowPipeSinkStatement extends ShowStatement implements IConfigStatement {
+import com.google.common.util.concurrent.ListenableFuture;
 
-  private String pipeSinkName;
+public class ShowPipeSinkTask implements IConfigTask {
 
-  public String getPipeSinkName() {
-    return pipeSinkName;
-  }
+  private ShowPipeSinkStatement showPipeSinkStatement;
 
-  public void setPipeSinkName(String pipeSinkName) {
-    this.pipeSinkName = pipeSinkName;
-  }
-
-  @Override
-  public QueryType getQueryType() {
-    return QueryType.READ;
+  public ShowPipeSinkTask(ShowPipeSinkStatement showPipeSinkStatement) {
+    this.showPipeSinkStatement = showPipeSinkStatement;
   }
 
   @Override
-  public <R, C> R accept(StatementVisitor<R, C> visitor, C context) {
-    return visitor.visitShowPipeSink(this, context);
+  public ListenableFuture<ConfigTaskResult> execute(IConfigTaskExecutor configTaskExecutor)
+      throws InterruptedException {
+    return configTaskExecutor.showPipeSink();
   }
 }
