+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/statement/sys/sync/CreatePipeSinkStatement.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.statement.sys;
+package org.apache.iotdb.db.mpp.plan.statement.sys.sync;
 
 import org.apache.iotdb.commons.path.PartialPath;
 import org.apache.iotdb.db.mpp.plan.analyze.QueryType;
@@ -28,21 +28,42 @@ import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
-public class StopPipeStatement extends Statement implements IConfigStatement {
+public class CreatePipeSinkStatement extends Statement implements IConfigStatement {
 
-  private String pipeName;
+  private String pipeSinkName;
 
-  public StopPipeStatement(StatementType stopPipeStatement) {
-    this.statementType = stopPipeStatement;
+  private String pipeSinkType;
+
+  private Map<String, String> attributes;
+
+  public CreatePipeSinkStatement(StatementType createPipeSinkStatement) {
+    this.statementType = createPipeSinkStatement;
   }
 
-  public String getPipeName() {
-    return pipeName;
+  public String getPipeSinkName() {
+    return pipeSinkName;
   }
 
-  public void setPipeName(String pipeName) {
-    this.pipeName = pipeName;
+  public Map<String, String> getAttributes() {
+    return attributes;
+  }
+
+  public String getPipeSinkType() {
+    return pipeSinkType;
+  }
+
+  public void setPipeSinkName(String pipeSinkName) {
+    this.pipeSinkName = pipeSinkName;
+  }
+
+  public void setPipeSinkType(String pipeSinkType) {
+    this.pipeSinkType = pipeSinkType;
+  }
+
+  public void setAttributes(Map<String, String> attributes) {
+    this.attributes = attributes;
   }
 
   @Override
@@ -57,6 +78,6 @@ public class StopPipeStatement extends Statement implements IConfigStatement {
 
   @Override
   public <R, C> R accept(StatementVisitor<R, C> visitor, C context) {
-    return visitor.visitStopPipe(this, context);
+    return visitor.visitCreatePipeSink(this, context);
   }
 }
