+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/statement/sys/sync/ShowPipeSinkTypeStatement.java
@@ -17,26 +17,15 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.execution.config.sys;
+package org.apache.iotdb.db.mpp.plan.statement.sys.sync;
 
-import org.apache.iotdb.db.mpp.plan.execution.config.ConfigTaskResult;
-import org.apache.iotdb.db.mpp.plan.execution.config.IConfigTask;
-import org.apache.iotdb.db.mpp.plan.execution.config.executor.IConfigTaskExecutor;
-import org.apache.iotdb.db.mpp.plan.statement.sys.StopPipeStatement;
+import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
+import org.apache.iotdb.db.mpp.plan.statement.metadata.ShowStatement;
 
-import com.google.common.util.concurrent.ListenableFuture;
-
-public class StopPipeTask implements IConfigTask {
-
-  private StopPipeStatement stopPipeStatement;
-
-  public StopPipeTask(StopPipeStatement stopPipeStatement) {
-    this.stopPipeStatement = stopPipeStatement;
-  }
+public class ShowPipeSinkTypeStatement extends ShowStatement {
 
   @Override
-  public ListenableFuture<ConfigTaskResult> execute(IConfigTaskExecutor configTaskExecutor)
-      throws InterruptedException {
-    return configTaskExecutor.stopPipe();
+  public <R, C> R accept(StatementVisitor<R, C> visitor, C context) {
+    return visitor.visitShowPipeSinkType(this, context);
   }
 }
