+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/statement/sys/sync/ShowPipeStatement.java
@@ -17,26 +17,34 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.execution.config.sys;
+package org.apache.iotdb.db.mpp.plan.statement.sys.sync;
 
-import org.apache.iotdb.db.mpp.plan.execution.config.ConfigTaskResult;
-import org.apache.iotdb.db.mpp.plan.execution.config.IConfigTask;
-import org.apache.iotdb.db.mpp.plan.execution.config.executor.IConfigTaskExecutor;
-import org.apache.iotdb.db.mpp.plan.statement.sys.ShowPipeSinkStatement;
+import org.apache.iotdb.db.mpp.plan.analyze.QueryType;
+import org.apache.iotdb.db.mpp.plan.statement.IConfigStatement;
+import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
+import org.apache.iotdb.db.mpp.plan.statement.metadata.ShowStatement;
 
-import com.google.common.util.concurrent.ListenableFuture;
+public class ShowPipeStatement extends ShowStatement implements IConfigStatement {
 
-public class ShowPipeSinkTask implements IConfigTask {
+  private String pipeName;
 
-  private ShowPipeSinkStatement showPipeSinkStatement;
+  public String getPipeName() {
+    return pipeName;
+  }
 
-  public ShowPipeSinkTask(ShowPipeSinkStatement showPipeSinkStatement) {
-    this.showPipeSinkStatement = showPipeSinkStatement;
+  public void setPipeName(String pipeName) {
+    this.pipeName = pipeName;
   }
 
   @Override
-  public ListenableFuture<ConfigTaskResult> execute(IConfigTaskExecutor configTaskExecutor)
-      throws InterruptedException {
-    return configTaskExecutor.showPipeSink();
+  public QueryType getQueryType() {
+    return QueryType.READ;
+  }
+
+  public ShowPipeStatement() {}
+
+  @Override
+  public <R, C> R accept(StatementVisitor<R, C> visitor, C context) {
+    return visitor.visitShowPipe(this, context);
   }
 }
