+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/statement/sys/sync/StartPipeStatement.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.statement.sys;
+package org.apache.iotdb.db.mpp.plan.statement.sys.sync;
 
 import org.apache.iotdb.commons.path.PartialPath;
 import org.apache.iotdb.db.mpp.plan.analyze.QueryType;
@@ -28,49 +28,21 @@ import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
-public class CreatePipeStatement extends Statement implements IConfigStatement {
+public class StartPipeStatement extends Statement implements IConfigStatement {
 
   private String pipeName;
-  private String pipeSinkName;
-  private long startTime;
-  private Map<String, String> pipeAttributes;
 
-  public CreatePipeStatement(StatementType createPipeStatement) {
-    this.statementType = createPipeStatement;
+  public StartPipeStatement(StatementType startPipeStatement) {
+    this.statementType = startPipeStatement;
   }
 
   public String getPipeName() {
     return pipeName;
   }
 
-  public String getPipeSinkName() {
-    return pipeSinkName;
-  }
-
-  public long getStartTime() {
-    return startTime;
-  }
-
-  public Map<String, String> getPipeAttributes() {
-    return pipeAttributes;
-  }
-
   public void setPipeName(String pipeName) {
     this.pipeName = pipeName;
-  }
-
-  public void setPipeSinkName(String pipeSinkName) {
-    this.pipeSinkName = pipeSinkName;
-  }
-
-  public void setStartTime(long startTime) {
-    this.startTime = startTime;
-  }
-
-  public void setPipeAttributes(Map<String, String> pipeAttributes) {
-    this.pipeAttributes = pipeAttributes;
   }
 
   @Override
@@ -85,6 +57,6 @@ public class CreatePipeStatement extends Statement implements IConfigStatement {
 
   @Override
   public <R, C> R accept(StatementVisitor<R, C> visitor, C context) {
-    return visitor.visitCreatePipe(this, context);
+    return visitor.visitStartPipe(this, context);
   }
 }
